+++ lib/api/api_calendar/api_calendar.dart
@@ -1,9 +1,11 @@
 import 'package:http/http.dart' as http;
 import 'package:template/screens/calendar_view.dart';
 import 'package:xml/xml.dart' as xml;
+import '../../models/model_claendarevents.dart';
 
 DateTime getDate = DateTime.now();
-String todaysDate = getDate.year.toString() + getDate.month.toString() + getDate.day.toString();
+String todaysDate =
+    getDate.year.toString() + getDate.month.toString() + getDate.day.toString();
 
 List<CalendarEvents?> parseXml(String xmlData) {
   var document = xml.XmlDocument.parse(xmlData);
@@ -38,9 +40,12 @@ List<CalendarEvents?> parseXml(String xmlData) {
 
 String processDescription(String description) {
   description = description.trim(); // Remove leading and trailing whitespace
-  description = description.replaceAll(r'\n', '\n'); // Convert '\n' to actual line breaks
-  description = description.replaceAll(RegExp(r'\n{1,2}$'), ''); // Remove one or two trailing newlines
-  description = description.replaceFirst(RegExp(r'^\n{1,2}'), ''); // Remove one or two leading newlines
+  description =
+      description.replaceAll(r'\n', '\n'); // Convert '\n' to actual line breaks
+  description = description.replaceAll(
+      RegExp(r'\n{1,2}$'), ''); // Remove one or two trailing newlines
+  description = description.replaceFirst(
+      RegExp(r'^\n{1,2}'), ''); // Remove one or two leading newlines
 
   return description;
 }
@@ -52,13 +57,16 @@ String processDecisionDate(String decisionDate) {
 
 Future<List<CalendarEvents>> getEvents() async {
   const url = 'https://data.riksdagen.se/kalender/?org=kamm&utformat=xml';
-  
+
   try {
     http.Response response = await http.get(Uri.parse(url));
     if (response.statusCode == 200) {
       List<CalendarEvents?> events = parseXml(response.body);
 
-      return events.where((event) => event != null).cast<CalendarEvents>().toList();
+      return events
+          .where((event) => event != null)
+          .cast<CalendarEvents>()
+          .toList();
     } else {
       print('Failed to load data: ${response.statusCode}');
       return [];
