+++ lib/api/api_homeview/api_home_view.dart
@@ -1,18 +1,20 @@
 import 'package:http/http.dart' as http;
 import 'dart:convert';
-import '../../screens/home_view.dart';
+import '../../models/model_dokument.dart';
 
 DateTime todaysDate = DateTime.now();
-DateTime startSearchDate = todaysDate.subtract(Duration(days:30 * 6));
+DateTime startSearchDate = todaysDate.subtract(Duration(days: 30 * 6));
 String numberofResults = '&sz=50';
 
-const String endPoint = 'https://data.riksdagen.se/dokumentlista/?sok=&doktyp=votering&rm=&ts=&bet=&tempbet=&nr=&org=&iid=&avd=&webbtv=&talare=&exakt=&planering=&facets=&sort=datum&sortorder=desc&rapport=&utformat=json&a=s';
+const String endPoint =
+    'https://data.riksdagen.se/dokumentlista/?sok=&doktyp=votering&rm=&ts=&bet=&tempbet=&nr=&org=&iid=&avd=&webbtv=&talare=&exakt=&planering=&facets=&sort=datum&sortorder=desc&rapport=&utformat=json&a=s';
 
-Future<List<Voteringar>> apiGetVotes() async {
-  String url = '$endPoint&from=$startSearchDate&tom=$todaysDate$numberofResults';
-  
+Future<List<Dokument>> apiGetDokuments() async {
+  String url =
+      '$endPoint&from=$startSearchDate&tom=$todaysDate$numberofResults';
+
   http.Response response = await http.get(Uri.parse(url));
   Map<String, dynamic> jsonData = jsonDecode(response.body);
-  List<dynamic> votesJson = jsonData['dokumentlista']['dokument'];
-  return votesJson.map((json) => Voteringar.fromJson(json)).toList();
+  List<dynamic> dokumentsJson = jsonData['dokumentlista']['dokument'];
+  return dokumentsJson.map((json) => Dokument.fromJson(json)).toList();
 }
