+++ lib/provider/provider_party_view.dart
@@ -1,7 +1,7 @@
 import 'package:template/api/api_party_view/api_party_view.dart';
 import 'package:flutter/material.dart';
 import 'package:template/provider/provider_infoview.dart';
-import 'package:template/api/api_infoview/api_single_votes.dart';
+import '.././api/api_get_title_from_xml.dart';
 
 import '../models/model_partyview_ledamot.dart';
 import '../models/model_partyview_ledamotresult.dart';
